refactor(web): extract Posts feature boundary
