+++ examples/web/posts.html
@@ -502,6 +502,6 @@
     </section>
   </main>
 
-  <script type="module" src="/src/post-app.ts"></script>
+  <script type="module" src="/src/entries/posts.ts"></script>
 </body>
 </html>
