+++ examples/web/src/entries/posts.ts
@@ -1,0 +1,3 @@
+import { mountPosts } from '../features/posts/browser/mount';
+
+mountPosts();
